requirements.txt
streamlit==1.2.0
pandas==1.2.4
numpy==1.20.1
tensorflow==2.7.0
tensorflow_hub==0.12.0
matplotlib==3.3.4
Pillow==8.2.0